--- frontend/app/components/request/type.ts
@@ -1,7 +1,0 @@
-import { RequestStatus } from "~/apiClient/apiSchemas";
-
-export type RequestStatusWithIdle = RequestStatus | "IDLE";
-export type RequestProps<
-  requestStatus extends RequestStatusWithIdle,
-  extraProps = NonNullable<object>,
-> = { status: requestStatus } & extraProps;
