+++ frontend/app/contexts/user-context.tsx
@@ -142,10 +142,7 @@ export const useClientAppRequest = (accessToken: string, id?: string) => {
       if (!m) return;
       return { latitude: Number(m[1]), longitude: Number(m[2]) };
     })();
-    const candidateAppRequest = { ...clientAppPayloadWithStatus };
-    if (debugStatus !== undefined) {
-      candidateAppRequest.status = debugStatus;
-    }
+    const candidateAppRequest = clientAppPayloadWithStatus;
     if (
       debugDestinationCoordinate &&
       candidateAppRequest?.payload?.coordinate
@@ -155,6 +152,7 @@ export const useClientAppRequest = (accessToken: string, id?: string) => {
     }
     return {
       ...candidateAppRequest,
+      status: debugStatus ?? candidateAppRequest?.status,
       auth: {
         accessToken,
       },
