+++ frontend/app/routes/client._index/route.tsx
@@ -1,9 +1,17 @@
 import type { MetaFunction } from "@remix-run/node";
-import type { FC } from "react";
-import { useClientAppRequestContext } from "../../contexts/user-context";
-import { Arrived } from "./requestComponent/arrived";
-import { Reception } from "./requestComponent/reception";
-import { Running } from "./requestComponent/running";
+import { useCallback, useRef, useState } from "react";
+import { fetchAppPostRequest } from "~/apiClient/apiComponents";
+import { Coordinate } from "~/apiClient/apiSchemas";
+import { useOnClickOutside } from "~/components/hooks/use-on-click-outside";
+import { LocationButton } from "~/components/modules/location-button/location-button";
+import { Map } from "~/components/modules/map/map";
+import { Button } from "~/components/primitives/button/button";
+import { Modal } from "~/components/primitives/modal/modal";
+import { Text } from "~/components/primitives/text/text";
+import { useClientAppRequestContext } from "~/contexts/user-context";
+import { Arrived } from "./driving-state/arrived";
+import { Carrying } from "./driving-state/carrying";
+import { Dispatched } from "./driving-state/dispatched";
 
 export const meta: MetaFunction = () => {
   return [
@@ -12,40 +20,133 @@ export const meta: MetaFunction = () => {
   ];
 };
 
-const ClientRequest: FC = () => {
-  const data = useClientAppRequestContext();
-  const requestStatus = data?.status ?? "IDLE";
-  const destinationCoordinate = data?.payload?.coordinate?.destination;
-  switch (requestStatus) {
-    case "IDLE":
-    case "MATCHING":
-    case "DISPATCHING":
-      return <Reception status={requestStatus} />;
-    case "DISPATCHED":
-      return (
-        <Running
-          status={requestStatus}
-          destinationCoordinate={destinationCoordinate}
-        />
-      );
-    case "CARRYING":
-      return (
-        <Running
-          status={requestStatus}
-          destinationCoordinate={destinationCoordinate}
-        />
-      );
-    case "ARRIVED":
-      return <Arrived />;
-    default:
-      return <div>unexpectedStatus: {requestStatus}</div>;
-  }
-};
+type Action = "from" | "to";
 
-export default function ClientRequestWrapper() {
+export default function Index() {
+  const data = useClientAppRequestContext();
+
+  const [action, setAction] = useState<Action>();
+  const [selectedLocation, setSelectedLocation] = useState<Coordinate>();
+  const [currentLocation, setCurrentLocation] = useState<Coordinate>();
+  const [destLocation, setDestLocation] = useState<Coordinate>();
+
+  const [isSelectorModalOpen, setIsSelectorModalOpen] = useState(false);
+  const selectorModalRef = useRef<HTMLElement & { close: () => void }>(null);
+  const handleSelectorModalClose = useCallback(() => {
+    if (selectorModalRef.current) {
+      selectorModalRef.current.close();
+    }
+  }, []);
+
+  const drivingStateModalRef = useRef(null);
+
+  const onClose = useCallback(() => {
+    if (action === "from") setCurrentLocation(selectedLocation);
+    if (action === "to") setDestLocation(selectedLocation);
+    setIsSelectorModalOpen(false);
+  }, [action, selectedLocation]);
+
+  const onMove = useCallback((coordinate: Coordinate) => {
+    setSelectedLocation(coordinate);
+  }, []);
+
+  const handleOpenModal = useCallback((action: Action) => {
+    setIsSelectorModalOpen(true);
+    setAction(action);
+  }, []);
+
+  // TODO: requestId をベースに配車キャンセルしたい
+  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
+  const [requestId, setRequestId] = useState<string>("");
+  const handleRideRequest = useCallback(async () => {
+    if (!currentLocation || !destLocation) {
+      return;
+    }
+    await fetchAppPostRequest({
+      body: {
+        pickup_coordinate: currentLocation,
+        destination_coordinate: destLocation,
+      },
+      headers: {
+        Authorization: `Bearer ${data.auth?.accessToken}`,
+      },
+    }).then((res) => setRequestId(res.request_id));
+  }, [data, currentLocation, destLocation]);
+
+  useOnClickOutside(selectorModalRef, handleSelectorModalClose);
+
   return (
     <>
-      <ClientRequest />
+      <Map
+        from={currentLocation}
+        to={destLocation}
+        initialCoordinate={selectedLocation}
+      />
+      <div className="w-full px-8 py-8 flex flex-col items-center justify-center">
+        <LocationButton
+          className="w-full"
+          location={currentLocation}
+          onClick={() => {
+            handleOpenModal("from");
+          }}
+          placeholder="現在地を選択する"
+          label="from"
+        />
+        <Text size="xl">↓</Text>
+        <LocationButton
+          location={destLocation}
+          className="w-full"
+          onClick={() => {
+            handleOpenModal("to");
+          }}
+          placeholder="目的地を選択する"
+          label="to"
+        />
+        <Button
+          variant="primary"
+          className="w-full mt-6 font-bold"
+          onClick={() => void handleRideRequest()}
+          disabled={!(Boolean(currentLocation) && Boolean(destLocation))}
+        >
+          ISURIDE
+        </Button>
+      </div>
+      {isSelectorModalOpen && (
+        <Modal ref={selectorModalRef} onClose={onClose}>
+          <div className="flex flex-col items-center mt-4 h-full">
+            <div className="flex-grow w-full max-h-[75%] mb-6">
+              <Map
+                onMove={onMove}
+                from={currentLocation}
+                to={destLocation}
+                initialCoordinate={
+                  action === "from" ? currentLocation : destLocation
+                }
+                selectable
+              />
+            </div>
+            <p className="font-bold mb-4 text-base">
+              {action === "from" ? "現在地" : "目的地"}を選択してください
+            </p>
+            <Button onClick={handleSelectorModalClose}>
+              {action === "from"
+                ? "この場所から移動する"
+                : "この場所に移動する"}
+            </Button>
+          </div>
+        </Modal>
+      )}
+      {data?.status && (
+        <Modal ref={drivingStateModalRef}>
+          {data.status === "DISPATCHED" && (
+            <Dispatched destLocation={data?.payload?.coordinate?.destination} />
+          )}
+          {data.status === "CARRYING" && (
+            <Carrying destLocation={data?.payload?.coordinate?.destination} />
+          )}
+          {data.status === "ARRIVED" && <Arrived />}
+        </Modal>
+      )}
     </>
   );
 }
