+++ frontend/app/routes/driver._index/driving-state/pickup.tsx
@@ -1,22 +1,14 @@
-import { useCallback } from "react";
+import { FC, useCallback } from "react";
 import { fetchChairPostRequestDepart } from "~/apiClient/apiComponents";
 import { CarGreenIcon } from "~/components/icon/car-green";
 import { CarRedIcon } from "~/components/icon/car-red";
 import { LocationButton } from "~/components/modules/location-button/location-button";
 import { Button } from "~/components/primitives/button/button";
 import { Text } from "~/components/primitives/text/text";
-import type { RequestProps } from "~/components/request/type";
 import { useClientChairRequestContext } from "~/contexts/driver-context";
-import type { ClientChairRequest } from "~/types";
 
-export const Pickup = ({
-  status,
-  payload,
-}: RequestProps<
-  "DISPATCHING" | "DISPATCHED" | "CARRYING",
-  { payload: ClientChairRequest["payload"] }
->) => {
-  const { auth } = useClientChairRequestContext();
+export const Pickup: FC = () => {
+  const { auth, payload, status } = useClientChairRequestContext();
 
   const handleDeparture = useCallback(async () => {
     await fetchChairPostRequestDepart({
@@ -32,7 +24,7 @@ export const Pickup = ({
   return (
     <>
       <div className="flex flex-col items-center my-8 gap-8">
-        {status === "DISPATCHING" ? (
+        {status === "DISPATCHING" && (
           <>
             <CarRedIcon className="size-[76px] mb-4" />
             <Text>
@@ -40,7 +32,8 @@ export const Pickup = ({
               さんの出発地点へ向かっています
             </Text>
           </>
-        ) : status === "DISPATCHED" ? (
+        )}
+        {status === "DISPATCHED" && (
           <>
             <CarGreenIcon className="size-[76px] mb-4" />
             <Text>
@@ -48,11 +41,6 @@ export const Pickup = ({
               さんの出発地点へ到着しました
             </Text>
           </>
-        ) : (
-          <Text>
-            <span className="font-bold mx-1">{payload?.user?.name}</span>
-            さんの配車依頼
-          </Text>
         )}
         <div className="flex flex-col w-full items-center px-8">
           <LocationButton label="from" disabled className="w-full" />
@@ -61,7 +49,7 @@ export const Pickup = ({
           <Text variant="danger" size="sm">
             到着予定時間: 21:58
           </Text>
-          {status === "DISPATCHED" ? (
+          {status === "DISPATCHED" && (
             <Button
               variant="primary"
               className="w-full mt-6"
@@ -69,7 +57,7 @@ export const Pickup = ({
             >
               出発
             </Button>
-          ) : null}
+          )}
         </div>
       </div>
     </>
